[ADD] folder name validation before App.loadSlideshow()
App now tries to validate the folder and shows an error if folder is not found.

test is done by trying to $.get() 0.html inside the given folder.
If get returns a success, the success callback is called (App.loadSlideshow(config).
If get returns a failure, the failure callback is called, showing the message to the user